Changed from int to double

+++ Point.cs
@@ -8,7 +8,7 @@ namespace ProjektarbeteV2
         public double Y { get; private set; }
         public double PointScore { get; private set; }
 
-        public Point(List<int> args) 
+        public Point(List<double> args) 
         {
             X = args[0];
             Y = args[1];
